Texture, music & terrain
I have tested some textures for the terrain (didnt work out), i added the correct textures for the green barrels (and changed their names to be more clear), and i put the music in for the water and tested that. (now it is a good size of 1 min and 15 sec).